Modified some debug output to show the specific offending code or name.
Modified how the packets are built for CIP writes.  If the entire request
fits in a single packet, then use the versions of the commands that are
not for fragments.  Reading boolean values turns out to fail if you
use fragment packets in some circumstances.  At least it worked when
I stopped using fragments.

Modified code that writes data from the tag to the request buffer on
CIP write such that it pads the data with a zero byte if the number
of bytes written is not a multiple of two.  Need to test if the
above change can be reverted with this.  It looks like not, but
more testing is needed.