Lots of small changes to make debugging out put selectable by a "debug" attribute in the tag attributes.   Set "debug=1" in the attributes to turn on debugging.  This has the unfortunately side effect of changing a lot of lines of code in small ways.  Testing is underway.
Because I forgot, there were some small changes to the ab_tag_cip_read_start()
routine to handle reading user defined data types and some default types
that are not 4 bytes in size such as STRING.  When reading these types,
the PLC will return data on boundaries of the element size, not on
boundaries of 4 bytes.  This has been tested with tags of STRING arrays,
UDTs of size 120 and 868.  In the last case, it turns out that the PLC
seems to break up the UDT right in the middle of an internal structure
without a problem.

Writes need this work still.