Free attributes memory after tag creation
In plc_tag_create, a call to attr_create_from_str allocates memory
that is passed to ab_tag_create, but then no longer used or referenced.
Added attr_destroy call after the return of ab_tag_create to free
memory.

Also, removed if-statement as it does not appear to do anything.