fix PLC5 access over DH+ bridge.  reorg code.
This commit fixes PLC5 access over a DH+ bridge (DHRIO only tested).

The code of the AB protocol was significantly reorganized due to
the requirements of the PLC-over-DH+ process.  That particular
type of access requires the use of connected messaging.  I cannot
find a way to do it with unconnected messages.  Perhaps there is one,
but it is not easy to find.

Correct Forward-Close messages will be sent when a connection-based
tag is deleted and there are no more using that connection object. Note
that if this is your only tag, the library will clean up too fast and
delete the session (and therefore the pending request) before it can
get sent.

There is a particular danger with PLC5 communication bridged over
DH+: it is really slow and it appears that either the DHRIO module
or something else in the system has very limited resources.  Creating
many tags and dropping them can result in Forward-Open (connection
setup command) failure.  These seem to come and go as if something
is cleaning up resources, but not fast enough.

So, keep the connection and read/write rates down.  Really down.

This is the first cut at the new 1.0 branch.  Once this is deemed
stable enough, it will be publicized.