Changed the order of tag deletion to the proper order for the test.
The order of the tag deletion should be 2 then 1.  This forces
the connection object to be destroyed and causes a Forward-Close
packet to be sent.  By delaying the destruction of the other tag,
we keep the session alive long enough to see the Forward-Close
being sent and the response (ignored) come back.