fixes for MicroLogix (old) and Micro 800.
There are two sets of fixes here.  The first is for the older Micrologix
family.  The encoding of data file names has changed from LAA to a level-based
scheme.  This works against PLC/5 systems at least.

The other (potential) fix is to change how packets are encoded for the newer
Micro 800 family.  These appear to require a null path argument as well, possibly,
as use of a non-fragmented read.  The latter needs more testing.