fixed pccc_encode_tag_name(): the endcoding for sub elements was not correct.
The encoding for sub elements had a while loop skipping past anything that
was not numeric.  This was not correct.  The only allowable trailing sub
element parts will be a '.' or a '/' followed by 'acc' or similar.  There
will be only one character as a delimiter and the part following will not
be numeric.