Fixes from Clang's static analyzer
In examples\multithread_plc5_dhp.c: this block will never be true as tag
is set to NULL in the while loop beforehand.

In lib\ab\eip_dhp_pccc.c: the deleted block is an exact duplicate of the
previous set of lines. Encoded name is being copied into data twice and
data is incremented twice incorrectly (correct me on this if I'm wrong)