Merged connection-based code from Micro850 branch, with cleanups.
Merged in the code from the Micro850 branch.  This adds connection-based
handling to both the Micro850 and LGX types.

Fixed code oops in plc_tag_destroy() that was returning out of a
critical_block area.

Added more pdebug output in the tag destroy call chain to track when
tags get unmapped.