changes to remove MS VS C compiler warnings
Mostly small changes to stop the compiler from warning about
casts etc.  Also changed the behavior of time_ms() to match
that in platform.c by returning a signed 64-bit integer
instead of unsigned.

Changed the name of sscanf so that it could be #defined to
sscanf on POSIX platforms and sscanf_s on Windows.  Note that
the latter is NOT completely API compatible!  For numbers, it
still works, but for strings and characters it does not.