added #define to work around sscanf warning for MS VS C compiler
MS VS does not like the old POSIX sscanf because it can overwrite memory
if you read in strings or characters (e.g. in a wide-character environment).
This hack just added a slight amount of indirection to allow Windows
to use sscanf_s and POSIX platforms to use sscanf as before.  All these
should be converted to safer use.