Changed to remove warnings on newer CMake version and adding .h files
Newer CMake versions do not like seeing quoted possible variables in
if() statements.  Unfortunately, one of the values that we need to check
for the type of compiler is the same as a variable name.  So, we explicitly set
the behavior to handle things the new way for this.

Added the .h files to the library and examples.  While CMake correctly set up
the dependencies at run time, it did not automatically add the .h files to the
MS Visual Studio project.  So, the .h files are manually included in the source
lists.