fixes for connected vs. unconnected messaging
This commit contains fixes for connected vs. unconnected messaging.
The old logic was incorrect.  However, the checking for whether a tag
needs a connection object or not is confusing now as part of it
occurs before the tag path is parsed and part of it occurs afterward.
This is all to use the single conn_path element in the tag.

This has been tested with LGX and PLC5 over DH+.