More refinements of handling of null paths.  Used for MLogix850.
The MicroLogix 850 uses CIP, sort of.  However, it has a couple of
quirks.  First, there is no internal routing path as in a ControlLogix
where you need to specify the path between the network module and the
CPU module.  Second, it must use connected messages.  In order to
support that, we need to suppress null path checks when the CPU
type is Micro800.   We also need to make sure that the path parsing
function can handle a null path and still insert the required
internal object path to handle the message.