Added ms output and added thread ID to debugging output.
The thread ID is more or less a fake. It it allocated per thread as a
small increasing integer.  It is for purposes of stitching the logs
together on a per-thread basis.  It is not the pthreads ID or the
tid of Linux.  THIS HAS NOT BEEN TESTED UNDER WINDOWS!

Also added millisecond output to the timestamp.  This should allow for
slightly better output handling.  Note that due to OS limitations, it is
likely that many lines will have the same ms value in the timestamp.