Changed output to factor out thread/time prefix, added line number and function to packet output.
This change factors out the prefix string which contains the thread ID and timestamp.  It adds
a prefix output to the packet output as well as outputting single lines at a time. This allows
the thread id to be used to filter logging messages including packets.  It also adds the
line number and function name to the packet output.