separate session creation into mutex locked and unlocked parts
This commit separates session creation into a part that happens with the global
mutex locked and a part that happens after.  The latter part now has the socket
creation and connection and the session and connection handshakes.

This has been tested for basic functionality, but needs more testing under high
load and multiple threads.

There are also some clean ups in the connection code to refactor a few bits into
lower parts of the call chain.