Updated debug output to have debug level, fixes for status loss when error on tag creation
This commit adds extra output in the debug output so that the debug level is printed in
the output.  It also has a number of small fixes for problems found during injected failure
testing during tag creation.  It also fixes a problem where the tag status was always set
to PLCTAG_STATUS_OK if the tag was created at all.