More clean up around request releasing
More clean ups to make sure that requests are released as soon as possible
in the main tag read/write routines.  Note that if the request was in
flight, we want to make sure that the IO thread cleans them up when
it sees them aborted.  We cannot do the clean up immediately in the read/write
routines as the IO thread might be doing some sort of manipulation of the
request at the time.