Changed API to provide a timeout for plc_tag_create and to make plc_tag an integer handle.
This includes two major changes.  The first is that the type of plc_tag changed from a
pointer to an integer.   The second is that plc_tag_create now takes a timeout and
behaves much like plc_tag_read.  Once key difference is that if tag creation fails,
the integer returned will be negative and will be the error.  Internally the tag will
be destroyed if anything went wrong.

This commit also includes updates to all the examples to use the new API.