Fixes for Windows DH+ bridge bug, issue #52.
This contains a fix for the Windows DH+ bridge problem.  There was an overwrite
due to the fact that str_copy was not correctly implemented on that platform.

This also includes some additional clean ups for request reference count handling
in sessions (on reception of a response, the request is removed from the session's
queue).

Some whitespace changes due to MSVC IDE.