Fixed limit bug in encoded tag name.
The original amount of space allocated for the encoded tag name was 64 bytes.   This is too short.
It turns out that AB seems to allow for 256 bytes.  So extend the encoded tag name space to 260 bytes for
some buffer.   Also extend the encoded path (could be longer?) to 200.   Could this be 256 too?

Bump the version number to denote the fix.