Adding ForwardOpenEx support to connection creation.
This code now tries to use ForwardOpenEx.   If that is not
supported, it tries ForwardOpen.   It tries a maximum size
for each type of open.   If it gets a failure due to the size,
it looks at the error response and tries to get the actual size
supported back and use that.