Adding new API functions.  Changed handling of API mutex.
This adds in plc_tag_create_sync() to take a timeout.

It changes how the tag API mutex is accessed.  Instead of an array of
mutexes per tag ID, this uses a single mutex allocated as part of the tag.

This is starting to add the new API functions for 64-bit integers and floats.