Adding Logix simulator to code base.
This is a _VERY_ rudimentary simulator.  It only supports two tags that
are hard coded to DINT types.   The tags are created in code only.   It takes
no command line arguments for tag types.

It only works with C11.   C99 is not supported.  It only works on Linux (or
at least it is only tested on Linux).