Added timeouts to sending and receiving packets.
These are zero in most places as we want the library client to
enforce any application-specific timeouts.   However, in the code
that sends and receives forward close messages, the messages are sent
during the destruction of the session and tag.   Therefore we need a
way to limit the time.   Forward close is a best-effort packet anyway.