Inserted casts to set types correctly.
This works around some warnings about integer sizes and makes the casts explicit.
Most of these give no warning when fewer warning flags are turned on.

These need to be eliminated over time.  However, there are a number of locations where
this will not be possible such as calculation of buffer lengths based on pointer
arithmetic that must be put into a uint16_t.