Back out addition of visible plc_tag_create_sync() API.
This is going to be the centerpiece of 2.0's visible changes.   Also,
needed to have an internal timeout because the semantics of the original
plc_tag_create() call changed.  It is possible to get a tag status of
PLCTAG_STATUS_PENDING now.  Before the session and connection were set up
at create time and so you would always get a tag that was either completely broken
or ready at the end of the create call.