Changed session thread handler logic and default timeouts for send/receive.
The session thread handler logic was turned into a state machine to allow for
controlled set up and tear down of sessions.  This refactored the handler routine
into a few more parts and cleaned up some of the session handling routines that
did not need to be "unsafe" anymore since the thread is the only caller.

Added default timeouts for all send and receiving of packets.   The logic changed a
bit to return those errors up the call chain.

Note that this now uses a default 2 second timeout for all basic operations in a
session including registration, forward open and general packet sending and receiving.