Changed code to add back support for ForwardOpenEx.
This also moved most of the constants involving packet size into the
session.c file and forced the use of the session variable max_payload_size
for size checking.

This still needs some work.  We should be using the request_capacity element
of the request in order to determine how much space we have to build a request.