Fixed bug with hashing: need to use the same salt!
There was a bug with calculating the hash.  The salt value I used was
the table pointer itself.  When calculating indexes when the table
was being expanded, the new table was used but then the arrays were
set back to the old table struct.  So the table pointer was different
and the wrong salt was used.

Added a new field in the hash table head struct to store the hash salt.
This is now used and the hash table now expands correctly.