Increased the limit of the number of tags in a bundle to 200.
The previous limit was 100 and even with normal tag names, this was easily hit and
only took up about 2200 bytes of the maximum 4000.   The increase did not change the
total time as much as I thought it work but that is probably due to the fact that
the test is only creating 150 tags.