Added check to clean up cached request references in process_requests.
If one of the processing steps failed, it looks like it is possible to leak
memory here by dropping some of the requests that were being bundled.

The new code checks if the return status is bad, if it is, it cleans up
all the requests that were being bundled.