Added 1ms sleep per iteration in stress_test.
It looks like slower machines or perhaps machines with bad threading will starve threads
and thus artifically fail.   This is particularly noticable when running under Valgrind.

This is not a problem in the library per se, but it does make it harder to test.   Must
make a note in the wiki about giving threads time.