Moved failed response check to check on CIP failure only in packed responses.
The check of the EIP header status stays in recv_eip_response() but the
check for whether the CIP status is OK is only done in process_requests() if
we are dealing with a bundled packet.  If not, then we punt out to the
originating tag.

This should be refactored so that session-specific errors are checked at
the session level.