Changed call to write() to use send() instead to block possible SIGPIPE.
If the remote system closes the TCP connection, we may throw a SIGPIPE on Linux
at the point, here, where we write to the socket.   This uses send() instead
and passes a flag to tell the system call not to throw signals.

Need to find a solution for BSD-like OSes like FreeBSD and macOS.