Fix reconnect errors.
This provides a fix for errors that occur when the remote PLC closes the connection.

The PLCs seem to time out after between 2 and 3 minutes.   At that point, we get errors
back from any writes to the socket.   We go into a close mode and report the error.  The
session will continue to try to reconnect after 5 seconds.