Fixed parsing problem for tags to allow multi-dimensional tag numbers to use commas. Cleaned up PLC type handling.
Multi-dimensional tags could not use notation like [0,10] for the array indexes.   This fixes that.

Code to determine the PLC type was duplicated in session.c and ab_common.c.  This was refactored into one core routine
in ab_common.c and used in both ab_common.c and session.c.