Fixed socket_close() issue.   Closes #120.
Thanks to user @srdgame.

When a socket was not completely open (i.e. non-existent IP address), the underlying fd was
set to 0. When socket_close() was called on Linux, this called close() on the fd, which closed fd 0!

This would close stdin on many systems or apparently an MQTT socket when using libmosquito!

The code for Windows looks like it was mostly immune to this as there was already a flag indicating
whether the socket was correctly opened or not. Under Linux, the flag was not being checked or set.