Fix for #123.
This is fairly invasive.   There are many small tweaks here to make sure that
any in-flight requests are freed when the library is shutting down.   More testing
is needed.   This has been tested with PLC/5, PLC/5 over DH+ and ControlLogix.

This introduces a new error code: PLCTAG_ERR_BUSY.

Here are the tweaks done here:

* Deeper, hopefully more complete checks to prevent multiple read or write operations
from being in flight at the same time for one tag.

* call the tag-specific abort function when plc_tag_delete is called.   This was not being
done, oddly.

* simplify the process that removes aborted requests from the session's queue.   It is now
a separate function and it called in multiple places.   Note that operations that inherently
hang a long time in the session, such as DNS lookup, will cause this clean up process to not
run

These seem to fix the two problems identified in #123: increasing queue depth when the connection
is dropped, and a memory leak when the library is shut down while a request is in flight.