Fix 188 - Modbus can only have one request in flight at a time per TCP connection. (#190)
* Force single request at a time in flight for Modbus.  Clean up dangling requests if the tag is gone.

* Bump version for Modbus fixes.

* Fixed RC bug. Cleanup after TCP shutdown.

There was a bug in the ref count handling when an existing PLC was reused in Modbus.   The rc_inc() call was not in place.

When an error occurs and the TCP connection is closed, we could be in a such a state that a request had been sent to the PLC but no response was returned.

In that case, we would never send another packet even if the TCP connection was reopened.   This fix clears the request_in_flight flag when the TCP connection is closed and there was a request already sent but no response.