v2.6.7 - Fixes for double free bug #546, changes to release process (#555)
This release contains a fix for bug
[#546](https://github.com/libplctag/libplctag/pull/546). Hopefully. It
is a Heisenbug that is difficult to trigger consistently.

Thanks to user "2gd4who" from the forum who noticed this.

There are also a lot of changes to how the releases are done. I broke up
the ZIP files that had contained everything for each target before. Now
each target has three ZIP files (example shown for macOS on ARM64):

* libplctag_2.6.7_macos_aarch64.zip - everything as before
* libplctag_2.6.7_macos_aarch64_tools.zip - the basic library and files
needed to develop with libplctag plus tools like tag_rw2, tag listing
tools etc.
* libplctag_2.6.7_macos_aarch64_dev.zip - just the files needed to
develop with libplctag.

The last two are experimental. Let me know if there is a flaw with them.

The final change is more automation around releases. This should help me
get them out the door more easily and more frequently. Now when a PR is
merged to the release branch, it automatically creates a release and
tags it.