v2.6.8 - Fix for #558 (#561)
The big on here is a fix for bug
[558](https://github.com/libplctag/libplctag/issues/558) found by user
@byunghyunna2. The largest changes were actually in the PCCC PLC code.
Please test well. All my tests pass, but I cannot cover all users
scenarios.

There are other fixes for building with Debug and sanitizers on. That
doesn't work, so the CMake config now prevents that.

Thanks to user dominoplayer84 for providing a test for parallel updates against
the PLC simulator.

The release can be found
[here](https://github.com/libplctag/libplctag/releases/tag/v2.6.8).

---------

Co-authored-by: Cedric Martineau <dominoplayer84@outlook.com>