Fix bug #565, error writing single bit in a tag
Thanks to user @mikecflood who found this. If you try to write a single
bit, it will fail with an error on CIP-based PLCs. This was something
that I broke with the rethinking of how the amount to write per packet
is calculated. Much more testing was added to my internal tests against
real hardware to make sure that this does not regress.

You can find the release in the [usual
place](https://github.com/libplctag/libplctag/releases/tag/v2.6.9).