v2.6.11 - More fixes for Micrologix/SLC bit writes, fix issue #572 (#576)
Thanks again to user @mikecflood for finding this. Writes to single bits
of SLC or Micrologix were not working correctly. Unfortunately, the
functions supported by the SLC/Micrologix PLCs do not allow for good RMW
(read-modify-write) support. Oddly, the PLC5 series has much better
support!

On a SLC/Micrologix PLC you can only write with a 16-bit mask. Due to
that I have limited the support for single bit writes (via bit tags) to
those tags that are 16 bits. This includes single elements from the B
and N files. You can read a bit tag from nearly anything, but not write.

I would really appreciate it if those of you who have SLC or Micrologix
PLCs hung off of DH+ that can be bridged from a ControlLogix could test
this. I do not have such a configuration so the code for that path is
untested.

This fixes issue #572.

You can find this release
[here](https://github.com/libplctag/libplctag/releases/tag/v2.6.11).