Start adding in reactor/buf/fsm code.   (#591)
This large commit adds much more detailed statistical analysis of fairness.  This drove significant changes in the Modbus code and some tweaks in the AB code.

A completely new implementation of the Modbus server is included that uses a new reactor-style core with buffer and state machine support code as well.   This has also be instrumented to make sure that latency is as low as possible.  This new core code will be the basis of any major refactoring of the EIP code later.